[cartel] Implement the Cartel class.
This class accepts multiple other actors as its constituents. It then
behaves as a single actor for all of these.

There is some magic here to preserve constraints. I found a solution
that works, but I need to think more about why it does.

The one part that is not implemented is the benefit-equalizing
algorithm, although if the constraint preservation works as well as it
appears to, this shouldn't be hard.